Some refactoring of the materials and items slots
